Combined IL2CPPUnityPlayer.cpp,  IL2CPPUnityPlayer.h, MonoUnityPlayer.cpp, and  MonoUnityPlayer.cpp into a unified UnityPlayer.cpp and UnityPlayer.h

+++ MelonLoader/MelonLoader.cpp
@@ -5,7 +5,6 @@
 #include "HookManager.h"
 #include "Logger.h"
 #include "ModHandler.h"
-#include "MonoUnityPlayer.h"
 
 bool MelonLoader::IsGameIl2Cpp = false;
 HINSTANCE MelonLoader::thisdll = NULL;
@@ -121,7 +120,7 @@ int MelonLoader::CountSubstring(std::string pat, std::string txt)
 	{
 		int j;
 		for (j = 0; j < M; j++)
-			if (txt[(long)(i) + j] != pat[j])
+			if (txt[i + j] != pat[j])
 				break;
 		if (j == M)
 		{
